VEM Problem Module

The vem_problem module defines the structure of VEM problems.

Available Types

Main.vem_problem.Mesh — Type

struct Mesh

Represents a mesh structure used in the Virtual Element Method (VEM).

Fields:

boundary::Vector{Int64}: A vector containing the indices of boundary vertices.

elements::Vector{Vector{Int64}}: A vector of vectors where each inner vector represents an element (e.g., triangle or polygon) by the indices of its vertices.

vertices::Vector{Tuple{Float64, Float64}}: A vector of tuples where each tuple represents the coordinates (x, y) of a vertex.

This struct is essential for defining the geometric domain in VEM computations.